# Add _prom_ext.re2_match(string, pattern) -> bool