--- a/Cargo.toml
+++ b/Cargo.toml
@@ -29,13 +29,16 @@ pg_test = ["serde_json", "proptest"]
 
 [dependencies]
 bincode = "1.3.1"
+lazy_static = "1.4.0"
+num_cpus = "1.13.1"
 pgx = "0.3.1"
 pgx-macros = "0.3.1"
-num_cpus = "1.13.1"
+proptest  = { version = "1.0.0", optional = true }
+regex = "1.5.6"
 sha2 = "0.10.2"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0.70", optional = true }
-proptest  = { version = "1.0.0", optional = true }
+uluru = "3.0.0"
 
 [build-dependencies]
 askama = "0.11.0"