## Description

This is the same as #362, but using a thread local `Rc` instead of a `Mutex`.
## Merge requirements

Please take into account the following non-code changes that you may need to make with your PR:

- [ ] CHANGELOG entry for user-facing changes
- [ ] Updated the relevant documentation